respectfully restore overview acquisition dialog box, shift 3dct from panel to a button in localization tab, fix imports and initialization
