+++ src/odemis/gui/cont/acquisition/overview_stream_acq.py
@@ -33,7 +33,7 @@ import wx
 
 from odemis.gui.preset import preset_as_is, get_global_settings_entries, \
     get_local_settings_entries, apply_preset
-from odemis.gui.win.acquisition import OverviewAcquisitionDialog
+from odemis.gui.win.acquisition import OverviewAcquisitionDialog, CorrelationDialog
 
 
 class OverviewStreamAcquiController(object):
@@ -101,3 +101,70 @@ class OverviewStreamAcquiController(object):
             return acq_dialog.data
         else:
             return None
+
+
+class CorrelationDialogController(object):
+    """Controller to handle the multi point correlation initialization"""
+
+    def __init__(self, tab_data, tab):
+        """
+        tab_data (MicroscopyGUIData): the representation of the microscope GUI
+        tab: (Tab): the tab which should show the data
+        """
+        self._tab_data_model = tab_data
+        self._main_data_model = tab_data.main
+        self._tab = tab
+
+    def open_correlation_dialog(self):
+        """
+        return None or a list of DataArrays: the acquired images. None if it was
+          cancelled.
+        """
+        # Indicate we are acquiring, especially important for the SEM which
+        # need to get the external signal to not scan (cf MicroscopeController)
+        self._main_data_model.is_acquiring.value = True
+
+        # save the original settings
+        settingsbar_controller = self._tab.settingsbar_controller
+        orig_entries = get_global_settings_entries(settingsbar_controller)
+        for sc in self._tab.streambar_controller.stream_controllers:
+            orig_entries += get_local_settings_entries(sc)
+        orig_settings = preset_as_is(orig_entries)
+        settingsbar_controller.pause()
+        settingsbar_controller.enable(False)
+
+        # pause all the live acquisitions
+        streambar_controller = self._tab.streambar_controller
+        streambar_controller.pauseStreams()
+        streambar_controller.pause()
+        streambar_controller.enable(False)
+
+        # create the dialog
+        try:
+            acq_dialog = CorrelationDialog(
+                self._tab.main_frame, self._tab_data_model)
+            parent_size = [v * 0.77 for v in self._tab.main_frame.GetSize()]
+
+            acq_dialog.SetSize(parent_size)
+            acq_dialog.Center()
+            action = acq_dialog.ShowModal()
+
+        except Exception:
+            logging.exception("Failed to create acquisition dialog")
+            raise
+        finally:
+            apply_preset(orig_settings)
+
+            settingsbar_controller.enable(True)
+            settingsbar_controller.resume()
+
+            streambar_controller.enable(True)
+            streambar_controller.resume()
+
+            self._main_data_model.is_acquiring.value = False
+            acq_dialog.Destroy()
+
+        # if action == wx.ID_OPEN:
+        #     return acq_dialog.data
+        # else:
+        #     return None
