+++ src/odemis/gui/model/__init__.py
@@ -28,7 +28,7 @@ from .stream_view import (ContentView, FeatureOverviewView, FeatureView,
                           FixedOverviewView, MicroscopeView, StreamView, View)
 from .tab_gui_data import (AcquisitionWindowData, ActuatorGUIData,
                            AnalysisGUIData, ChamberGUIData, CryoChamberGUIData,
-                           CryoCorrelationGUIData, CryoGUIData,
+                           CryoCorrelationGUIData, CryoTdctCorrelationGUIData, CryoGUIData,
                            CryoLocalizationGUIData, EnzelAlignGUIData,
                            FastEMAcquisitionGUIData, FastEMMainTabGUIData,
                            FastEMSetupGUIData, LiveViewGUIData, MicroscopyGUIData,
